[SQL] Remove MapIndex operators after tables with primary keys
