+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/frontend/parser/SqlDropObject.java
@@ -1,7 +1,6 @@
 package org.dbsp.sqlCompiler.compiler.frontend.parser;
 
 import com.google.common.collect.ImmutableList;
-import org.apache.calcite.jdbc.CalcitePrepare;
 import org.apache.calcite.sql.SqlDrop;
 import org.apache.calcite.sql.SqlIdentifier;
 import org.apache.calcite.sql.SqlNode;
