+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/postgres/PostgresLimitTests.java
@@ -3,7 +3,6 @@ package org.dbsp.sqlCompiler.compiler.sql.postgres;
 import org.dbsp.sqlCompiler.compiler.CompilerOptions;
 import org.dbsp.sqlCompiler.compiler.DBSPCompiler;
 import org.dbsp.sqlCompiler.compiler.sql.tools.SqlIoTest;
-import org.junit.Ignore;
 import org.junit.Test;
 
 // Adapted from https://github.com/postgres/postgres/src/test/regress/expected/limit.out
