+++ sql-to-dbsp-compiler/simulator/src/main/java/org/dbsp/simulator/values/IntegerSqlValue.java
@@ -1,7 +1,6 @@
 package org.dbsp.simulator.values;
 
 import org.dbsp.simulator.types.IntegerSqlType;
-import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nullable;
 import java.util.Comparator;
